Re-structured public  screens folder structure
