+++ src/components/screens/public/public-routes.tsx
@@ -1,8 +1,8 @@
 import ConnectionHandler from "components/contexts/connection-handler";
 import React from "react";
 import { Route, Routes } from "react-router-dom";
-import FormScreen from "../forms/form-screen";
-import PublicFormsScreen from "./public-forms-screen";
+import PublicFormScreen from "./public-form-screen";
+import PublicWelcomeScreen from "./public-welcome-screen";
 
 /**
  * Component for public routes
@@ -11,11 +11,11 @@ const PublicRoutes: React.FC = () => (
   <Routes>
     <Route
       path="/"
-      element={ <PublicFormsScreen/> }
+      element={ <PublicWelcomeScreen/> }
     />
     <Route
       path="/:metaformSlug"
-      element={ <ConnectionHandler><FormScreen/></ConnectionHandler> }
+      element={ <ConnectionHandler><PublicFormScreen/></ConnectionHandler> }
     />
   </Routes>
 );
